requirements.txt
aiogram==3.13.1
aiohttp==3.10.11
aiofiles==24.1.0
aiosqlite==0.20.0
apscheduler==3.10.4
python-dotenv==1.0.1
